drivers: msm_thermal_simple: Fix format specifier warning

+++ drivers/thermal/qcom/msm_thermal_simple.c
@@ -104,7 +104,8 @@ static void thermal_throttle_worker(struct work_struct *work)
 
 	/* Update thermal zone if it changed */
 	if (new_zone != old_zone) {
-		pr_info("temp_avg: %i, batt: %i, cpus: %i\n", temp_avg, temp_batt, temp_cpus_avg);
+		pr_info("temp_avg: %lli, batt: %lli, cpus: %lli\n", 
+			(int long long)temp_avg, (int long long)temp_batt, (int long long)temp_cpus_avg);
 		t->curr_zone = new_zone;
 	}
 	update_online_cpu_policy();
